@@ -1,7 +1,9 @@
-const express = require('express');
-const router = express.Router();
-const { handleAgentQuery } = require('../controllers/agentController');
+import express from 'express';
+import { handleAgentQuery } from '../controllers/agentController.js';
 
+const router = express.Router();
+
+// This will be accessible at /api/ask
 router.post('/ask', handleAgentQuery);
 
-module.exports = router;
+export default router;
